+++ resources/js/components/ChestCard.vue
@@ -24,7 +24,7 @@
                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
                     <a class="dropdown-item" :href="chest.permalink"><i class="fas fa-link fa-fw mr-1"></i> {{ __('Permalink') }}</a>
                     <h6 class="dropdown-header" v-if="chest.editable">{{ __('Manage') }}</h6>
-                    <a class="dropdown-item" :href="chest.url_edit" v-if="chest.editable"><i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}</a>
+                    <a class="dropdown-item" @click="edit = chest" v-if="chest.editable"><i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}</a>
                     <confirm class="dropdown-item"
                              :text="`<i class='fas fa-trash-alt fa-fw mr-1'></i> ${__('Delete')}`"
                              :text-confirm="`<i class='fas fa-check fa-fw mr-1'></i> ${__('Confirm')}`"
@@ -34,6 +34,12 @@
                 </div>
             </div>
         </div>
+
+        <modal :open="edit !== false" @closed="edit = false" size="xl">
+            <template #content>
+                <chest-form :chest="edit" class="mb-0"></chest-form>
+            </template>
+        </modal>
     </div>
 
 </template>
@@ -51,6 +57,12 @@ export default {
             required: true,
             default: () => {}
         },
+    },
+
+    data() {
+        return {
+            edit: false,
+        }
     },
 
     methods: {