+++ resources/js/components/StoryCard.vue
@@ -24,7 +24,7 @@
                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
                     <a class="dropdown-item" :href="story.url"><i class="fas fa-link fa-fw mr-1"></i> {{ __('Permalink') }}</a>
                     <h6 class="dropdown-header" v-if="story.editable">{{ __('Manage') }}</h6>
-                    <a class="dropdown-item" :href="story.url_edit" v-if="story.editable"><i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}</a>
+                    <a class="dropdown-item" @click="edit = story" v-if="story.editable"><i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}</a>
                     <confirm class="dropdown-item"
                              :text="`<i class='fas fa-trash-alt fa-fw mr-1'></i> ${__('Delete')}`"
                              :text-confirm="`<i class='fas fa-check fa-fw mr-1'></i> ${__('Confirm')}`"
@@ -34,8 +34,13 @@
                 </div>
             </div>
         </div>
-    </div>
 
+        <modal :open="edit !== false" @closed="edit = false" size="xl">
+            <template #content>
+                <story-form :story="edit" class="mb-0"></story-form>
+            </template>
+        </modal>
+    </div>
 </template>
 
 <script>
@@ -51,6 +56,12 @@ export default {
             required: true,
             default: () => {}
         },
+    },
+
+    data() {
+        return {
+            edit: false,
+        }
     },
 
     methods: {